+++ core-primitives/attestation-handler/Cargo.toml
@@ -1,0 +1,62 @@
+[package]
+name = "itp-attestation-handler"
+version = "0.8.0"
+authors = ["Integritee AG <hello@integritee.network>"]
+edition = "2021"
+
+[target.'cfg(not(target_env = "sgx"))'.dependencies]
+sgx_tse = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+sgx_tstd = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git", features = ["untrusted_fs","net","backtrace"] }
+sgx_rand = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+sgx_types = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+sgx_tcrypto = { branch = "master", git = "https://github.com/apache/teaclave-sgx-sdk.git" }
+
+[dependencies]
+codec  = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+arrayvec = { version = "0.7.1", default-features = false }
+hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
+thiserror = { version = "1.0", optional = true }
+thiserror_sgx = { package = "thiserror", git = "https://github.com/mesalock-linux/thiserror-sgx", tag = "sgx_1.1.3", optional = true }
+
+# scs / integritee
+itp-utils = { default-features = false, features = ["sgx"], git = "https://github.com/integritee-network/pallets.git", branch = "master" }
+substrate-api-client = { default-features = false, git = "https://github.com/scs/substrate-api-client.git", branch = "polkadot-v0.9.27" }
+
+# mesalock
+webpki = { git = "https://github.com/mesalock-linux/webpki", branch = "mesalock_sgx" }
+webpki-roots = { git = "https://github.com/mesalock-linux/webpki-roots", branch = "mesalock_sgx" }
+log = { git = "https://github.com/integritee-network/log-sgx" }
+serde_json = { tag = "sgx_1.1.3", git = "https://github.com/mesalock-linux/serde-json-sgx" }
+yasna = { rev = "sgx_1.1.3", default-features = false, features = ["bit-vec", "num-bigint", "chrono", "mesalock_sgx"], git = "https://github.com/mesalock-linux/yasna.rs-sgx" }
+rustls = { rev = "sgx_1.1.3", features = ["dangerous_configuration"], git = "https://github.com/mesalock-linux/rustls" }
+
+# for attestation
+chrono = { git = "https://github.com/mesalock-linux/chrono-sgx" }
+base64 = { rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/rust-base64-sgx" }
+num-bigint = { git = "https://github.com/mesalock-linux/num-bigint-sgx" }
+httparse = { default-features = false, features = ["mesalock_sgx"], git = "https://github.com/integritee-network/httparse-sgx", branch = "sgx-experimental" }
+itertools = { version = "0.10.1",  default-features = false, features = []}
+bit-vec = { version = "0.6",    default-features = false }
+
+# local deps
+itp-node-api = { path = "../node-api", default-features = false, features = ["sgx"] }
+itp-ocall-api = { path = "../ocall-api", default-features = false }
+itp-settings = { path = "../settings" }
+itp-sgx-externalities = { path = "../substrate-sgx/externalities", default-features = false, features = ["sgx"] }
+itp-sgx-io = { path = "../sgx/io", default-features = false, features = ["sgx"] }
+itp-sgx-crypto = { path = "../sgx/crypto", default-features = false, features = ["sgx"] }
+
+# substrate deps
+sp-core = { default-features = false, features = ["full_crypto"], git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27" }
+
+[patch.crates-io]
+log = { git = "https://github.com/integritee-network/log-sgx" }
+
+[patch."https://github.com/mesalock-linux/log-sgx"]
+log = { git = "https://github.com/integritee-network/log-sgx" }
+
+[features]
+default = ["sgx"]
+std = ["thiserror"]
+sgx = ["thiserror_sgx"]
+test = []
