+++ core-primitives/attestation-handler/src/error.rs
@@ -1,0 +1,57 @@
+/*
+	Copyright 2021 Integritee AG and Supercomputing Systems AG
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+		http://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+
+*/
+
+#[cfg(all(not(feature = "std"), feature = "sgx"))]
+use crate::sgx_reexport_prelude::*;
+use sgx_tstd as std;
+use sgx_tstd::boxed::Box;
+use sgx_types::sgx_status_t;
+
+pub type Result<T> = core::result::Result<T, Error>;
+
+/// Parentchain block importer error.
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+	#[error("SGX error, status: {0}")]
+	Sgx(sgx_status_t),
+	#[error(transparent)]
+	Other(#[from] Box<dyn std::error::Error + Sync + Send + 'static>),
+	#[error("{0}")]
+	IoError(#[from] std::io::Error),
+	#[error("Crypto error: {0}")]
+	Crypto(itp_sgx_crypto::Error),
+}
+
+impl From<sgx_status_t> for Error {
+	fn from(sgx_status: sgx_status_t) -> Self {
+		Self::Sgx(sgx_status)
+	}
+}
+
+impl From<itp_sgx_crypto::error::Error> for Error {
+	fn from(error: itp_sgx_crypto::error::Error) -> Self {
+		Self::Crypto(error)
+	}
+}
+
+impl From<Error> for sgx_status_t {
+	/// return sgx_status for top level enclave functions
+	fn from(error: Error) -> sgx_status_t {
+		log::warn!("Transform non-sgx-error into `SGX_ERROR_UNEXPECTED`: {:?}", error);
+		sgx_status_t::SGX_ERROR_UNEXPECTED
+	}
+}
