+++ enclave-runtime/src/test/tests_main.rs
@@ -18,7 +18,7 @@
 use crate::test::evm_pallet_tests;
 
 use crate::{
-	attestation, rpc,
+	rpc,
 	sync::tests::{enclave_rw_lock_works, sidechain_rw_lock_works},
 	test::{
 		cert_tests::*,
@@ -75,7 +75,7 @@ use std::{string::String, sync::Arc, time::Duration, vec::Vec};
 #[no_mangle]
 pub extern "C" fn test_main_entrance() -> size_t {
 	rsgx_unit_tests!(
-		attestation::tests::decode_spid_works,
+		itp_attestation_handler::attestation_handler::tests::decode_spid_works,
 		stf_sgx_tests::enclave_account_initialization_works,
 		stf_sgx_tests::shield_funds_increments_signer_account_nonce,
 		stf_sgx_tests::test_root_account_exists_after_initialization,
